+++ DrawingApp/DrawingApp/Point.swift
@@ -1,0 +1,13 @@
+//
+//  Point.swift
+//  DrawingApp
+//
+//  Created by jsj on 2022/03/03.
+//
+
+import Foundation
+
+struct Point {
+    let x: Double
+    let y: Double
+}
