+++ DrawingApp/DrawingApp/RectangleFactory.swift
@@ -14,9 +14,9 @@ enum ShapeFactoryType {
     func make() -> Rectangle {
         switch self {
         case .rectangle(let width, let height):
-            return RectangleFactory.make(size: Rectangle.Size(width: width, height: height))
+            return RectangleFactory.make(size: Size(width: width, height: height))
         case .square(let length):
-            return RectangleFactory.make(size: Rectangle.Size(width: length, height: length))
+            return RectangleFactory.make(size: Size(width: length, height: length))
         }
     }
 }
@@ -26,8 +26,8 @@ final class RectangleFactory {
     static func make(colorType: Color.Standard, x: Double = 0, y: Double, width: Double = 100, height: Double = 100) -> Rectangle {
         let uid = UniqueID.generate()
         return Rectangle(uid: uid,
-                         point: Rectangle.Point(x: x, y: y),
-                         size: Rectangle.Size(width: width, height: height),
+                         point: Point(x: x, y: y),
+                         size: Size(width: width, height: height),
                          color: colorType.rgb,
                          alpha: 1)
     }
@@ -35,21 +35,21 @@ final class RectangleFactory {
     static func make(x: Double, y: Double, width: Double, height: Double) -> Rectangle {
         let uid = UniqueID.generate()
         return Rectangle(uid: uid,
-                         point: Rectangle.Point(x: x, y: y),
-                         size: Rectangle.Size(width: width, height: height),
+                         point: Point(x: x, y: y),
+                         size: Size(width: width, height: height),
                          color: Color.Standard.white.rgb, alpha: 1)
     }
     
     static func make(x: Int, y: Int, width: Int, height: Int) -> Rectangle {
         let uid = UniqueID.generate()
         return Rectangle(uid: uid,
-                         point: Rectangle.Point(x: Double(x), y: Double(y)),
-                         size: Rectangle.Size(width: Double(width), height: Double(height)),
+                         point: Point(x: Double(x), y: Double(y)),
+                         size: Size(width: Double(width), height: Double(height)),
                          color: Color.Standard.white.rgb,
                          alpha: 1)
     }
     
-    static func make(origin: Rectangle.Point = .init(x: 0, y: 0), size: Rectangle.Size) -> Rectangle {
+    static func make(origin: Point = Point(x: 0, y: 0), size: Size) -> Rectangle {
         let uid = UniqueID.generate()
         return Rectangle(uid: uid,
                          point: origin,
