+++ DrawingApp/DrawingApp/Size.swift
@@ -1,0 +1,13 @@
+//
+//  Size.swift
+//  DrawingApp
+//
+//  Created by jsj on 2022/03/03.
+//
+
+import Foundation
+
+struct Size {
+    let width: Double
+    let height: Double
+}
